requirements.txt
httpx>=0.27.0
tenacity>=8.2.0
supabase==2.10.0
python-dotenv>=1.0.0
anthropic>=0.84.0
pydantic>=2.0.0
google-genai>=1.0.0
Pillow>=10.0.0
youtube-transcript-api>=1.0.0
